fastapi-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
ultralytics==8.3.34
pillow==10.4.0
numpy==2.1.2
python-multipart==0.0.12